implemented ConsoleController (without testing)
restructured ProcessRunnerPackage
