+++ .gitignore
@@ -1,4 +1,4 @@
 /modules/
 /scripts/
 /ProcessSignaling.iml
-/src/main/java/TestingGrounds/
+/src/main/java/testingGrounds/
