--- src/main/java/Logger/LogType.java
@@ -1,20 +1,0 @@
-package Logger;
-
-/**
- * log type enum
- */
-public enum LogType {
-    DEBUG("[DEBUG]"),
-    ERROR("[ERROR]"),
-    INFO("[INFO]"),
-    MESSAGE("[MESSAGE]"),
-    WARNING("[WARNING]");
-
-    /**
-     * constructor
-     * @param s log type
-     */
-    LogType(String s) {
-
-    }
-}
