+++ src/main/java/logger/LogType.java
@@ -1,0 +1,20 @@
+package logger;
+
+/**
+ * log type enum
+ */
+public enum LogType {
+    DEBUG("[DEBUG]"),
+    ERROR("[ERROR]"),
+    INFO("[INFO]"),
+    MESSAGE("[MESSAGE]"),
+    WARNING("[WARNING]");
+
+    /**
+     * constructor
+     * @param s log type
+     */
+    LogType(String s) {
+
+    }
+}
