--- src/main/java/processManagement/ProcessExceptions/ProcessAlreadyStartedException.java
@@ -1,15 +1,0 @@
-package processManagement.ProcessExceptions;
-
-public class ProcessAlreadyStartedException extends Throwable {
-    public ProcessAlreadyStartedException() {
-        super("process is already alive");
-    }
-
-    public ProcessAlreadyStartedException(String message) {
-        super(message);
-    }
-
-    public ProcessAlreadyStartedException(Throwable cause) {
-        super(cause);
-    }
-}
