--- src/main/java/processManagement/ProcessExceptions/ProcessCouldNotStopException.java
@@ -1,15 +1,0 @@
-package processManagement.ProcessExceptions;
-
-public class ProcessCouldNotStopException extends Throwable {
-    public ProcessCouldNotStopException() {
-        super("called process wasn't able to stop");
-    }
-
-    public ProcessCouldNotStopException(String message) {
-        super(message);
-    }
-
-    public ProcessCouldNotStopException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
