--- src/main/java/processManagement/ProcessRunner/ProcessRunnerType.java
@@ -1,10 +1,0 @@
-package processManagement.ProcessRunner;
-
-public enum ProcessRunnerType {
-    STANDARD_RUNNER,
-    DIRECT_COMMAND_RUNNER,
-    SOCKET_COMMUNICATION_RUNNER,
-    PRECOMPILE_RUNNER,
-    SCRIPT_RUNNER,
-    CUSTOM
-}
