+++ src/main/java/processManagement/processExceptions/InterpreterOrScriptNotDefinedException.java
@@ -1,0 +1,15 @@
+package processManagement.processExceptions;
+
+public class InterpreterOrScriptNotDefinedException extends Throwable {
+    public InterpreterOrScriptNotDefinedException() {
+        super("no interpreter or script was given or it was empty");
+    }
+
+    public InterpreterOrScriptNotDefinedException(String message) {
+        super(message);
+    }
+
+    public InterpreterOrScriptNotDefinedException(Throwable cause) {
+        super(cause);
+    }
+}
