+++ src/main/java/processManagement/processExceptions/ProcessCouldNotBeReproducedException.java
@@ -1,0 +1,15 @@
+package processManagement.processExceptions;
+
+public class ProcessCouldNotBeReproducedException extends Throwable {
+    public ProcessCouldNotBeReproducedException() {
+        super("couldn't reproduce Process");
+    }
+
+    public ProcessCouldNotBeReproducedException(String message) {
+        super(message);
+    }
+
+    public ProcessCouldNotBeReproducedException(Throwable cause) {
+        super(cause);
+    }
+}
