+++ src/main/java/processManagement/processExceptions/ProcessNotExitedYetException.java
@@ -1,0 +1,15 @@
+package processManagement.processExceptions;
+
+public class ProcessNotExitedYetException extends Exception {
+    public ProcessNotExitedYetException() {
+        super("process is still alive");
+    }
+
+    public ProcessNotExitedYetException(String message) {
+        super(message);
+    }
+
+    public ProcessNotExitedYetException(Throwable cause) {
+        super(cause);
+    }
+}
